+++ SharpSource/SharpSource.Test/Helpers/CSCodeFix+Test.cs
@@ -40,6 +40,12 @@ public static partial class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
             TestState.AdditionalReferences.Add(typeof(Microsoft.VisualStudio.TestTools.UnitTesting.TestClassAttribute).Assembly.Location);
             TestState.AdditionalReferences.Add(typeof(Newtonsoft.Json.JsonSerializer).Assembly.Location);
             TestState.AdditionalReferences.Add(typeof(Microsoft.Extensions.Logging.LoggerMessageAttribute).Assembly.Location);
+            TestState.AdditionalReferences.Add(typeof(Elastic.Clients.Elasticsearch.ElasticsearchClient).Assembly.Location);
+            TestState.AdditionalReferences.Add(typeof(Elastic.Transport.TransportResponse).Assembly.Location);
+            TestState.AdditionalReferences.Add(typeof(Nest.ElasticClient).Assembly.Location);
+            TestState.AdditionalReferences.Add(typeof(Elasticsearch.Net.IElasticsearchResponse).Assembly.Location);
+            TestState.AdditionalReferences.Add(typeof(OpenSearch.Client.OpenSearchClient).Assembly.Location);
+            TestState.AdditionalReferences.Add(typeof(OpenSearch.Net.IOpenSearchResponse).Assembly.Location);
 
             // Initialized explicitly so the underlying test framework doesn't auto-inject all the netcoreapp3.1 references
             // Unfortunately MS stopped updating the utility library that abstracted this: https://github.com/dotnet/roslyn-sdk/issues/1047