+++ SharpSource/SharpSource.Test/RecursiveEqualityOperatorOverloadTests.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-
+using SharpSource.Test.Helpers;
 using VerifyCS = SharpSource.Test.CSharpCodeFixVerifier<SharpSource.Diagnostics.RecursiveOperatorOverloadAnalyzer, Microsoft.CodeAnalysis.Testing.EmptyCodeFixProvider>;
 
 namespace SharpSource.Test;
@@ -75,6 +75,30 @@ namespace ConsoleApplication1
 		    var a = 1 == 1;
 		    return a1;
 	    }
+    }
+}";
+
+        await VerifyCS.VerifyNoDiagnostic(original);
+    }
+
+    [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/341")]
+    public async Task RecursiveEqualityOperatorOverload_WithDifferentOverload()
+    {
+        var original = @"
+using System.Collections.Generic;
+using System.Linq;
+
+namespace ConsoleApplication1
+{
+    file record Range(int Lower, int Upper)
+    {
+        public static IEnumerable<Range> operator -(Range minuend, Range subtrahend)
+        {
+            yield return minuend;
+        }
+
+        public static IEnumerable<Range> operator -(IEnumerable<Range> minuends, Range subtrahend)
+            => minuends.SelectMany(minuend => minuend - subtrahend);
     }
 }";